fix: allow multiple URI schemes during discovery
This commit addresses a potential problem which could arise
if the scheme of a URI obtained via CoRE Resource Discovery
was not the same as the URI scheme of the original discovery URI.
This was caused by the fact that only one ProtocolClient for
the original URI scheme was initialized.

This commit addresses the issue by introducing a cache for clients
which is expanded if more clients for different URI schemes are needed
during discovery. In the stop method, all clients are cleaned up
simultaneously.